Support v2 SHOW FUNCTIONS
